refactor: 💡 Restructure to prevent packages from aliasing
