+++ internal/provider/datasources/slackappmanifest/features.go
@@ -11,7 +11,7 @@ import (
 	"github.com/hashicorp/terraform-plugin-framework/schema/validator"
 	"github.com/hashicorp/terraform-plugin-framework/types"
 
-	listvalidatorextensions "github.com/yumemi-inc/terraform-provider-slackapp/internal/extensions/listvalidator"
+	"github.com/yumemi-inc/terraform-provider-slackapp/internal/listvalidatorx"
 	"github.com/yumemi-inc/terraform-provider-slackapp/internal/slack/manifest"
 	"github.com/yumemi-inc/terraform-provider-slackapp/internal/typeconv"
 )
@@ -130,7 +130,7 @@ func (*Shortcut) schema() *schema.ListNestedBlock {
 			},
 		},
 		Validators: []validator.List{
-			listvalidatorextensions.SizeAtMostWarning(5),
+			listvalidatorx.SizeAtMostWarning(5),
 		},
 	}
 }
@@ -190,7 +190,7 @@ func (*SlashCommand) schema() *schema.ListNestedBlock {
 			},
 		},
 		Validators: []validator.List{
-			listvalidatorextensions.SizeAtMostWarning(5),
+			listvalidatorx.SizeAtMostWarning(5),
 		},
 	}
 }
